feat(it01hu01): AppRouter implemented
