+++ src/Auth/pages/LoginPage.tsx
@@ -1,0 +1,6 @@
+
+export const LoginPage = () => {
+    return(
+        <h1>LoginPage</h1>
+    )
+}
