+++ src/main.tsx
@@ -2,7 +2,8 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import {App} from './App'
 
-import mapboxgl from 'mapbox-gl'; // or "const mapboxgl = require('mapbox-gl');"
+import mapboxgl from 'mapbox-gl';
+import {BrowserRouter} from "react-router-dom";
 
 // const { VITE_MAP_API_KEY } = getEnvVariables();
 
@@ -10,6 +11,8 @@ mapboxgl.accessToken = 'pk.eyJ1IjoibXItcGF5dSIsImEiOiJjbG54bzlnNzIwa3p4MmxuMHNyM
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-      <App />
+      <BrowserRouter>
+        <App />
+      </BrowserRouter>
   </React.StrictMode>,
 )
